+++ header/light/Base/inline/Buffer.inl
@@ -2,16 +2,16 @@
 
 namespace light
 {
-    template <class Type, u32 Size>
+    template <class Type, u64 Size>
     Type&
-    Buffer<Type, Size>::operator[](u32 index)
+    Buffer<Type, Size>::operator[](u64 index)
     {
         return data[index];
     }
 
-    template <class Type, u32 Size>
+    template <class Type, u64 Size>
     const Type&
-    Buffer<Type, Size>::operator[](u32 index) const
+    Buffer<Type, Size>::operator[](u64 index) const
     {
         return data[index];
     }
